Adds prerequisite to serial communication to VHF
This file allows for changing the VHF board between CDC and Hybrid
communication.

In CDC mode, the board can be accessed with PySerial, in the same
way minicom does communicate with it. This allows for buffer flushing,
ADCINIT, CLOCKINIT etc.

After these are done, one can then set back to Hybrid mode for data
collection.