Adds FIFO clear methods
teststream.exec and vhf_board.softlink are symlinks that point to the
compiled teststream executable and the VHF board as located under /dev
respectively. These symlinks will be referenced by all Python files
that interact with the VHF board.

set_device_mode.cpp has updated to allow for udev-adm reloading.

toggle_bConfigValue is a python file that acts as remotely removing and
reconnecting the board to have it show up in the relevant /dev/serial/
folder.

clear_fifo is a Python file that can be runned once that should clear
the FIFO light as expected.