Adds USBDDS python controller
This commit provides a comparably more fine grained control of the
USBDDS board in ways not yet done on the USBDDS repository. However,
this commit does not aim to achieve the arbitary waveform as desired
in that other repository.

This python wrapper is provided as an alternative to the function
generator to provision the testing of the unwrapped phase in context
of the local reference frame of the RF oscillator.