Adds recording of pure sine signals
For testing the demodulation scheme, taking a pre-determined sine signal
of known frequency relative to the sampling frequency allows us to think
of the signal as increasing/decreasing in phase(time) by considering its
relation to the 'origin' in the 'detector' reference frame. This commit
provisions a few subfolder scripts that saves the stream output into a
convenient location for assisting with the analysis of the demodulation
function.