Adds teststream wrapper with run_vhf.py
This commit provisions the base run_vhf.py that has been modified to
have obtained `vhf_*/run_vhf_with_*.py` for those corresponding niche
cases. This code is intended as a general wrapper for teststream, but is
not necessarily feature complete yet, for example, with variables still
being coded directly into the .py script currently.

The code features some CLI to feedback to the user the configuration
which teststream will be invoked with, along with some debugging and
expected timeout.