Utilise Obspy 1.2.2 defaults
Obspy 1.2.2 had different defaults for wlen (and possibly others) in
plotting of spectrogram. A rigorous pixel comparision between a local
install of obspy 1.2.2 and obspy 1.4.0 utilising obspy 1.2.2 defaults
is yet to be done.
Nonetheless, the state of the script is therefore much closer to the
original plot (prior to Words contrast edit) created by Miguel.