Rewrite figure generating code
traffic-spectrum.py is a full rewrite and optimization with similar
logic to Miguel's original FFT-VHF.py, but with the build up of
knowledge through the internals available. This commit serves as a
foundation for getting a txt output to pass into GNU plot eventually.

traffic-spectrum.py has yet to crop the array's freq axis and Sxx to
give a proper cmap for the relevant plot of Fig 37 (accounting that
a large vertical line saturates the upper end of L* in the plot).