Add cmap.FuncNorm for specifying own colour curve
The choice of c_forward is motivated by defining a "compact supported"
version of a sigmoid curve, which gives rise to the use of the
"sigmoid" curve as defined on Wiki:Bump Function. We then scale this
S-curve ourself specifically to the dataset (i.e.: Sxx) that has been
generated here, i.e.: we wish for largest variation in colourspace at
z = some z_0.

Manually specifing the inverse function (as required by FuncNorm) is
non-trivial, and so, we just utilise pynverse's inversefunc.

However, we then note that automated testing of c_forward means that we
cannot have an entirely flat f(x) outside the compact support, and so,
we arbitarily add a tiny linear function (but sufficiently large) to
pass the automated testing of inverse retrieval. (It will become
self-apparent when gradient in the linear term is too small.)

Current parameters in this commit are intended to the specific dataset
under creation. Edit as necessary.

Additional code is specific to generating the expected Fig size for
LaTeX.