Add phase plot
gridspec method is introduced for clean alignment of phase axis
compared to the nex ax generated from fig.colorbar
With the inclusion of cmap truncate method, it is now possible for the
colorbar to only show the relevant segment of the colorbar to the
occluded imshow. Combining with vmap, it is now possible to recreate a
cleaner version of Miguel's spectrogram.