Add file selection via GUI
Using xdg-desktop-portals, kdialog and QtFileDialog are some GUI methods
to select files. We put them into their own functions and call them in
succession if the prior fails, and fall back into the previous text
selection menu method, with several features that are left to be desired
for said method.
For convenience, relevant modules are specified in requirements.txt too.