Fix improper main() in plot_VHF_output
README includes some necessary dependencies for PyGObject.
We note that the XDG-Desktop-Portal is likely to open on the primary
screen instead of forwarding over SSH. Either investigation into GTK to
see if use of $DISPLAY is possible, or if to provide a user flag to
bypass into KDialog instead has to be done.