Provide check for if VHF board is presently under use
Through the use of fuser, with the assumption that only the current
logged in user is orchestrating the experiment, we force
subprocess_run to quit the python program if in strict_use is enabled
(this is made default first), so as to avoid having multiple processes
trying to interface with the VHF board at once. This however is yet to
stop clear_FIFO from doing anything dangerous.