Introduce child process wrapper in the context of VHF pool management
Added a few tests to make sure proper instantization and closure of
child processes so that Pool has an easier job selecting a ready job for
root process to work with.

Added pytest to perform unit-testing, with test for IdentifiedProcess
class to ensure it behaves more or less to expectation.