Add request_requeue signal from genericVHF
Not only does genericVHF now signal when VHF is complete, it will also
send a request_requeue signal to signal that the after-trace analysis is
completed. This is useful particularly for long analysis where root
process might mistakenly assume genericVHF is prepared to perform a
subsequent sample. It is still root's responsibility to init with
sufficient child processes to ensure that no genericVHF child processes
are not available.

Also fixed improper use of abstract_property during previous commit move
of main_func to genericVHF.