Fix corruption of data over Pipe between processes
File Descriptors are leaked during os.fork, we close the child end of
the pipe in the parent process after fork. This seems to have fixed
corruption of serde binary data across the pipe.

Correspondingly, IdentifiedProcess no longer requires the initializer
have the child end of the Pipe as an attribute. All dependencies are
correspondingly propagated.