Add default channel for function generator
With proper type hinting as to "my_resource" attribute, it was
discovered that the overloading of write and query methods should have
been independent of the serial interface method. This is rectified along
with default channel being part of class initialization, with a name
change to emphasise on that the class was designed for single channel
operation in mind.