Add root/pool managed init example using Function Generator to VHF
test_Pool_integrate_IdentifiedProcess is the main script being invoked
here, with the VHF board being connected to the Function Generator. So
far, the script only tests for initialization time to be successful. It
has been noted that broken pipes and EOF occur during KeyboardInterrupt,
and so are thrown as an exception and their observed failure points.
This might not necessarily be the correct way to solve the issue at
hand. It is also noted that Child Processes still are killed without
being found in either alive or zombie.
Otherwise, the provided scripts serves as the init-time foundation
necessary prior to adding looping of the independent variable (function
generator) prior to the subsequent round of VHF sampling.