Adds test for child process handling failure in init of IdentifiedProcess
This test writes a few mocks in an attempt to test if child processes
exit gracefully when the Parent process's end of the pipe closes as the
child process initializes. Despite the fact that in a python shell,
closing one end of a Pipe() and trying to send data from the other end
invokes such an error, trying to close the pipe using a different thread
in the test during the blocking step, that is the creation of the
IdentifiedProcess, could not be done, and so the test is not a good test
of behaviour as to how the ideal child should exit.