Change logging location for func_gen tester
The log file generated was created relative to the position of the root
python code that was the parent process involved in testing the radius
as a function of Vpp driven into the ADC board, and so, during the
promotion of the code out of vhf_func_gen/test/... into vhf_func_gen/,
it was unintentionally not accounted for. The corresponding name of the
log file generated has also been changed to better reflect the role of
the code.