Add manifold rollover tracking
A class named ManifoldManager is added in this commit, where it
represents the state of the entire trace file having manifold rollovers.
To ensure responsivity, we only consume blocks of the trace file at a
time, and update the class's internals in the update method. Only after
consuming the entire file do we lock this class in preparation for it to
return to the user the relevant m offset necessary.