Inform VHFParser of plot window in initializer
This commit lets VHFParser prepare its internals for the desired plot
window as given in an argument to the initialization process. However,
manifold fixing has yet to be implemented, and is therefore removed in
this commit; and also is the ability of a created instance of VHFParser
to update the plot window yet to be done.