Add unit-test for spectrogram crop
It has became apparent that imshow having origin="lower" or "upper" was
making all the difference into the cropping behaviour it takes to get
the appropriate frequency selection. This commit therefore is a unit
test to show that the problem comes from not understanding imshow, and
that there is no issue for spectrogram crop.