Add Chirp-Z transform spectrogram as referenced against mlab
With reference to mlab's spectrogram, we create our own variant of
spectrogram for a specific use case (obtaining Power Spectrum Density),
for our choice of frequency range.
Does not support other options at the moment, such as magnitude etc. In
the future, a version written against Scipy's ShortTimeFFT is
preferable; but it must be noted that not creating the entire frequency
range via Chirp-Z means that inverse STFT is not possible.