Fix parser for higher sampling frequencies
For sampling frequencies that were so large such that 1/f < 1ms,
timedelta as provided from standard library would have just truncated
out the value, and the corresponding timedelta object would have just
been of the 0 interval.
Resolving this by artificially scaling TraceTimer's internal
representation using microseconds as natural units of only sampling
interval was not viable, as it would have then rounded off the plot
window's duration count to the floor(?) of the scaling parameter, but
might be possible if everything else was also scaled accordingly.
Regardless, the quick fix is to utilize something that extends the
standard library's timedelta with nanosecond support, with increased
runtime checks for butchering of the sampling interval variable.