Add multiple binary digestor for Spectrogram average in 15-min intervals
There's quite a lot of code smells (Timestamp etc, that will become very
apparent during the data unpacking for plotting) to be had in the code
being committed, but is currently just being placed into the history for
future references.

The core logic behind data_to_npy (more accurately, npz files), is being
able to select either a single file, or multiple files as being part of
a single 15-min window for which we are obtaining the Spectrogram
average on (across all frequencies at once). There's some painful logic
for selecting across arbitrary number of files if they fit into the
15-min window. For the purpose of a standardised poster against Miguel's
plot currently, binary VHF traces are being downsampled to a
standardised frequency of 40 kHz first before applying whatever nonsense
procedure of getting the Spectrogram per 2-min slice. All sliding
windows accumulating up to the end of the end of a 15 min interval
contributes to one data point for plotting.

Fluffier logic does this in random order to avoid collision on the same
.bin file due to multiple threads doing more or less the same few
timings within the same few successions.