Add plotters for SAB2024 and QE-II figures
Here, we read the npz file as provided by the prior `data_to_npy.py`
file, and plot out the result in a manner fitting for posters.
As such, some attempt to ensure good axes and shapes are placed into the
code.

Hardcoded if statements and axes formatters were done by Justin are
specific to getting the plotted figure to look nice, and is data
specific. More work needs to be done to handle automatic axes ticks
formatting.