Add plotters for individual poster
It has became apparent that mangling the SAB_plot.py script to provide
the necessary figures for the SAB-Individual Project posters to be too
much of a hassle, as the shape of the figure is just fundamentally
different, with a different narrative being aimed for too. We point to
the data we want and just create the plot directly instead.