Update binary digestor to account for Pandas removal
Prior to this commit, this branch creating SAB 2024 plots were based on
0.5.0.2. This commit prepares for a rebase onto 0.5.1, which contains
the proper fix of TraceTimer that does not involve Panda's Timestamp.

This commit is an attempt to clean up some of the code smells that was
a consequence of the Timestamp type being leaked out of TraceTimer's
internals, but the propagation of the type has fallen short
approximately around line 112 involving update_plot_timings, which is
unable to consume np.datetime64. The binary digestor in the next commit
should instead consider for mutating the output of np.arange back into
datetime instead of retaining as np.datetime64, which is the bulk of
this commit.

In the longer future, a proper library method to handle multi-file
handling instead is required.