Use naive datetimes for digestor
As mentioned in the prior commit, the removal of Pandas has made it abit
easier to reason about the available methods as the type of object
(datetime/np.datetime64/~~pandas.Timestamp~~). This commit imposes
timings_left() to return naive datetime.datetime (in contrast to the
previous commit's failed attempt setting it np.datetime64). However,
there's still a bit of redundant aware_to_naive being done to interact
with the fact that the entire infrastructure has to now work around the
user-desired data point being (timezone-)naive.