Use aware datetimes for digestor and plotter
As mentioned in a prior commit about having timings_left() return
naive-datetimes creating some redundancies, we now have timings_left()
return (timezone-)aware datetimes, and this consequently makes it a lot
easier to handle the interaction with VHFparser. We only call for
naivety when having to interact with numpy arrays specifically for
passing to matplotlib still.