Fix timings_left in data_to_npy
During the rewrite to utilise aware datetimes from the beginning as much
as possible, the timings_left function was not properly checked to
ensure setdiff1d was able to accommodate the 2 different data types.
This commit fixes it, and so multiple invocations of data_to_npy should
not amount to a full re-run again.